+++ app/Services/PluginMachineApi.php
@@ -8,6 +8,7 @@ use Illuminate\Support\ServiceProvider;
 use Illuminate\Support\Facades\Http;
 use Illuminate\Support\Arr;
 use Illuminate\Http\Client\Response;
+use Log;
 
 /**
  * API client for Plugin Machine